==> 03_Ref_genes_download.ipynb <==
# Download all flower development related gene sequences
### Created 3 sets of gene-seqs:
* Set of genes from **3x-filtered** flower development related PMIDs: **108 genes** - saved in `refgenes/two_times_filtered`
* Set of genes from **2x-filtered** flower development related PMIDs: **165 genes** - saved in `refgenes/three_times_filtered`
* (Additional) Set of genes from **earlier transcriptomic papers**, where they were proved to be DEG in flower development related experiments: this list is not comparable to the other two, as the genes havent been proved to have a real flower-related function! It is only interesting to look at such lists and re-analyse them. The used papers: ..., The number of genes: **~170 genes** - saved in `refgenes/DEG_analyses_based`
## 1. Data import

library(taxize)
library(dplyr,  warn.conflicts = FALSE)
library(genbankr)
library(rentrez)
library(usethis)
library(parallel)
library(foreach, warn.conflicts = FALSE)
library(doParallel)
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refpapers/")

# pmid-flower-list with 2x-overlap - 898 PMIDs
pmid_flower <- readRDS("PMIDlist_flower.rds")
pmid_flower <- as.integer(pmid_flower)
pmid_flower <- as.data.frame(pmid_flower, rownames = FALSE, stringsAsFactors = FALSE)
colnames(pmid_flower) <- "pmid"
glimpse(pmid_flower)
# pmid-flower-list with 3x-overlap - 308 PMIDs
pmid_flower_three <- readRDS("PMIDlist_flower_3.rds")
pmid_flower_three <- as.integer(pmid_flower_three)
pmid_flower_three <- as.data.frame(pmid_flower_three, rownames = FALSE, stringsAsFactors = FALSE)
colnames(pmid_flower_three) <- "pmid"
glimpse(pmid_flower_three)

# Downloaded Gene2Pubmed from ftp://ftp.ncbi.nih.gov/gene/DATA/gene2pubmed.gz @03.04.2020
gene2pm <- read.table("gene2pubmed", sep = "")
colnames(gene2pm) <- c("taxid", "geneid", "pmid")
glimpse(gene2pm)
head(gene2pm)

## 2. Get *Triticeae*-taxIDs into single vector

uid <- get_uid(sciname = "Triticeae")
uid[1]

# 147389 is the id for Triticeae
tax_children <- downstream(uid[1], downto = "species", db = "ncbi")
glimpse(tax_children)

**For future problem solving:**
I had to put `ENTREZ_KEY='532056952c2098c0cd03a43bc25e345a7f08'` into my `.Renviron` file in the path of `/home/pgsb/vanda.marosi/` using the `usethis::edit_r_environ()` and terminal navigation with bash and restarted R

# convert to data.frame and data wrangling for tidy data
attributes(tax_children) <- NULL
tax <- as.data.frame(tax_children, stringsAsFactors = FALSE)
colnames(tax) <- c("taxid", "name", "rank")
glimpse(tax)

# total of 389 taxid-s found for triticeae
tax_id <- as.integer(tax$taxid)
taxid <- as.data.frame(tax_id, rownames = FALSE)
colnames(taxid) <- "taxid"
str(taxid)

## 3. Intersect `tax_id` and `pmid_flower` with `Gene2Pubmed` table

#intersect taxid with gene2pubmed
trittax_geneid_pmid <- inner_join(gene2pm, taxid, by = "taxid")
glimpse(trittax_geneid_pmid)
# 27 930 geneIDs are linked to Triticeae taxid-s

# intersect pmid_flower with gene2pubmed that was already filtered with taxid-s to get only Triticeae relevant genes
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refgenes")
# for 2x-overlap list 165 genes are found
trittax_geneid_flowerpmid <- inner_join(trittax_geneid_pmid, pmid_flower, by = "pmid")
write.table(trittax_geneid_flowerpmid, file = "triticeae_twoxfiltered_geneid_pmid.txt", append = FALSE, quote = FALSE, sep = "\t", dec = ".",
            row.names = FALSE, col.names = TRUE)

# save list for Uniport analyses
two <- select(trittax_geneid_flowerpmid, geneid)
glimpse(two)
write.table(two, file = "triticeae_two_genelist.txt", append = FALSE, quote = FALSE, sep = "\t", dec = ".",
            row.names = FALSE, col.names = TRUE)
glimpse(trittax_geneid_flowerpmid)

# for 3x-overlap list 108 genes are found - focusing on this list mainly!!
trittax_geneid_flowerpmid_three <- inner_join(trittax_geneid_pmid, pmid_flower_three, by = "pmid")

# save list for Uniport analyses
three <- select(trittax_geneid_flowerpmid_three, geneid)
glimpse(three)
write.table(three, file = "triticeae_three_genelist.txt", append = FALSE, quote = FALSE, sep = "\t", dec = ".",
            row.names = FALSE, col.names = TRUE)
glimpse(trittax_geneid_flowerpmid_three)

# 3.1 Inspect few example transcriptomics article whether it is present with genes in `Gene2pubmed` database/table
* title: Identification of Wheat Inflorescence Development-Related Genes Using a Comparative Transcriptomics Approach - PMID 29581960 
* title: Genome-Wide Investigation of Heat Shock Transcription Factor Family in Wheat (Triticum aestivum L.) and Possible Roles in Anther Development. - PMID 31963482

#trittax_geneid_flowerpmid$pmid[[29581960]]
#which(trittax_geneid_flowerpmid$pmid == 29581960)
#which(pmid_flower == 29581960)
which(gene2pm$pmid == 29581960)
which(gene2pm$pmid == 31963482)
# they are not present on the GeneID list

# 4. Download Geneseqs for flower-dev related geneIDs
### 4.1 `Rentrez`
* for batch download from `nuccore` database both in the formats of `rettype =`
    - **Genbank**:`gb` = GenBank flat file OR `gbwithparts` = GenBank flat file with full sequence (contigs)
    - **FASTA**: `fasta_cds_na` = CDS nucleotide FASTA OR `fasta_cds_aa` = CDS protein FASTA (for nucleotide I just used `fasta` bc the other did not give results for all the id-s)

### 4.1.1 `Three_times_filtered` flower-dev related list of genes

# create genelist from dataframe
genelist <- trittax_geneid_flowerpmid_three$geneid
tail(genelist)

# to fetch a single GeneID I have to link from GeneDataBase to Nuccore (works with different UID)
# within $link choose $gene_nuccore_refseqrna (for transcript) but most cases just have $gene_nuccore and the two numbers are equal in a few checked examples
# here is an example how to access id-conversion
gene_ids <- c(109738555)
linked_seq_ids <- entrez_link(dbfrom="gene", id=gene_ids, db="nuccore")
glimpse(linked_seq_ids)
linked_seq_ids$links

nuccore_id_list <- foreach(i=genelist, .packages ="rentrez", .combine = 'c') %do% (entrez_link(dbfrom="gene", id = i, db="nuccore"))$links$gene_nuccore 
str(nuccore_id_list)

# loop for Genbank format
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refgenes/three_times_filtered/genbank")
cl <- makeCluster(3)
registerDoParallel(cl)
a <- foreach(i=nuccore_id_list, .packages="rentrez") %dopar% {
    seq <- entrez_fetch(db="nuccore", id=i, rettype="gbwithparts")
    write(seq, file=paste(i,".gb",sep=""))
    }
stopCluster(cl)

# differences between data-types
# nucleotide CDS
seq1 <- entrez_fetch(db="nuccore", id=298704707, rettype="fasta_cds_na")
class(seq1)
nchar(seq1)
cat(strwrap(substr(seq1, 1, 200)), sep="\n")
# protein CDS
seq2 <- entrez_fetch(db="nuccore", id=298704707, rettype="fasta_cds_aa")
nchar(seq2)
cat(strwrap(substr(seq2, 1, 200)), sep="\n")
# fasta
seq3 <- entrez_fetch(db="nuccore", id=298704707, rettype="fasta")
nchar(seq3)
cat(strwrap(substr(seq3, 1, 200)), sep="\n")
# GenBank flat file
seq4 <- entrez_fetch(db="nuccore", id=298704707, rettype="gb")
nchar(seq4)
cat(strwrap(substr(seq4, 1, 300)), sep="\n")
# GenBank flat file with full sequence
seq5 <- entrez_fetch(db="nuccore", id=298704707, rettype="gbwithparts")
nchar(seq5)
cat(strwrap(substr(seq5, 1, 300)), sep="\n")

# loop for fasta_nucleotide format into fasta (fasta_cds_na doesnt give results for all the id-s)
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refgenes/three_times_filtered/fasta_nucleotide")
cl <- makeCluster(3)
registerDoParallel(cl)
b <- foreach(i=nuccore_id_list, .packages="rentrez") %dopar% {
    seq <- entrez_fetch(db="nuccore", id=i, rettype="fasta")
    write(seq, file=paste(i,".fasta",sep=""))
    }
stopCluster(cl)

# loop for fasta_protein format into - doesnt seem to work for all! empty ones need to be deleted in bash
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refgenes/three_times_filtered/fasta_protein")
cl <- makeCluster(3)
registerDoParallel(cl)
c <- foreach(i=nuccore_id_list, .packages="rentrez") %dopar% {
    seq <- entrez_fetch(db="nuccore", id=i, rettype="fasta_cds_aa")
    write(seq, file=paste(i,".fasta",sep=""))
    }
stopCluster(cl)

### 4.1.1 `Two_times_filtered` flower-dev related list of genes

# create genelist from dataframe
genelist_two <- trittax_geneid_flowerpmid$geneid
glimpse(genelist_two)
# convert to nuccore id-s
nuccore_id_list_two <- foreach(i=genelist_two, .packages ="rentrez", .combine = 'c') %do% (entrez_link(dbfrom="gene", id = i, db="nuccore"))$links$gene_nuccore 

setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refgenes/two_times_filtered")
write.table(nuccore_id_list_two, file = "twotimes_filtered_accession.txt", append = FALSE, quote = FALSE, sep = "\t", dec = ".",
            row.names = FALSE, col.names = TRUE)
glimpse(nuccore_id_list_two)

# loop for Genbank format
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refgenes/two_times_filtered/genbank")
cl <- makeCluster(3)
registerDoParallel(cl)
a <- foreach(i=nuccore_id_list_two, .packages="rentrez") %dopar% {
    seq <- entrez_fetch(db="nuccore", id=i, rettype="gbwithparts")
    write(seq, file=paste(i,".gb",sep=""))
    }
stopCluster(cl)

# loop for fasta_nucleotide format into fasta (fasta_cds_na doesnt give results for all the id-s)
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refgenes/two_times_filtered/fasta_nucleotide")
cl <- makeCluster(3)
registerDoParallel(cl)
b <- foreach(i=nuccore_id_list_two, .packages="rentrez") %dopar% {
    seq <- entrez_fetch(db="nuccore", id=i, rettype="fasta")
    write(seq, file=paste(i,".fasta",sep=""))
    }
stopCluster(cl)

#  loop for fasta_protein format into - doesnt seem to work for all! empty ones need to be deleted in bash
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refgenes/two_times_filtered/fasta_protein")
cl <- makeCluster(3)
registerDoParallel(cl)
c <- foreach(i=nuccore_id_list_two, .packages="rentrez") %dopar% {
    seq <- entrez_fetch(db="nuccore", id=i, rettype="fasta_cds_aa")
    write(seq, file=paste(i,".fasta",sep=""))
    }
stopCluster(cl)

### 4.1.1 Additionally collected flower-dev related list of genes

Daniel sent one, + the first paper he sent, + the papers that came from transcriptomics filter but were experimental

would be interesting to create a set as "false negatives" but after finishing the rest of the analysis

# 5. Mapping
## 5.1 Parsing sequence files with `Genbank-r`
* a package for parsing GenBank files into semantically useful objects, able to read Genbank files
* `readGenBank()`: takes either local file or text in a form of a GBAccession object containing Nuccore versioned accession numbers
    - uses `rentrez`
    - generates a GenBankFull object, which contains the annotations and the origin sequence
    - useful tutorial: https://bioconductor.org/packages/release/bioc/vignettes/genbankr/inst/doc/genbankr.html

Sys.getenv("ENTREZ_KEY")

gb_names <- list.files(path="/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refgenes/two_times_filtered/genbank", full.names=FALSE)
str(gb_names)

setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refgenes/two_times_filtered/genbank")
suppressPackageStartupMessages(library(genbankr))
#here is a package default example
#smpfile = system.file("sample.gbk", package="genbankr")
#here is one of my files
gb = parseGenBank("1027405096.gb")

str(gb)

#extract information for grouping files: organism, mol_type
a <- gb$FEATURES
a <- unlist(a)
a[[6]]
a[[7]]

#extract metadata into vector
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refgenes/two_times_filtered/genbank")
gb_organism <- foreach(i=gb_names, .packages ="genbankr", .combine = "c") %do% (suppressWarnings(unlist(parseGenBank(i)$FEATURES))[[6]])
gb_moltype <- foreach(i=gb_names, .packages ="genbankr", .combine = "c") %do% (suppressWarnings(unlist(parseGenBank(i)$FEATURES))[[7]])
glimpse(gb_organism)
glimpse(gb_moltype)

gb_names <- as.data.frame(gb_names, stringsAsFactors = FALSE)
gb_organism <- as.data.frame(gb_organism, stringsAsFactors = FALSE)
gb_moltype <- as.data.frame(gb_moltype, stringsAsFactors = FALSE)
gb_meta <- bind_cols(gb_names, gb_organism, gb_moltype)
colnames(gb_meta) <- c("filename", "organism", "molecule")
glimpse(gb_meta)

### Separation of fasta_nucleotide files based on organism & mol_type

library(stringr)
# grouping by organism
triticum <- dplyr::filter(gb_meta, organism == "Triticum aestivum")
glimpse(triticum)
hordeum <- dplyr::filter(gb_meta, organism %in% c("Hordeum vulgare","Hordeum vulgare subsp. vulgare"))
glimpse(hordeum)
aegilops <- dplyr::filter(gb_meta, organism %in% c("Aegilops tauschii","Aegilops tauschii subsp. tauschii"))
glimpse(aegilops)

# group triticum by molecule
# nosplicing
t_nosplicing <- dplyr::filter(triticum, molecule %in% c("genomic DNA"))
glimpse(t_nosplicing)
# save filename-list
t_nosplicing_filename <- pull(t_nosplicing, filename)
t_nosplicing_filename <- str_replace(t_nosplicing_filename, ".gb", ".fasta")
glimpse(t_nosplicing_filename)
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refgenes/gmap/triticum/")
write.table(t_nosplicing_filename, file = "wheat_nosplicing.csv", append = FALSE, quote = FALSE, sep = " ", dec = ".",
            row.names = FALSE, col.names = FALSE)

# splicing
t_splicing <- dplyr::filter(triticum, molecule %in% c("unassigned DNA", "transcribed RNA", "mRNA"))
glimpse(t_splicing)
# save filename-list
t_splicing_filename <- pull(t_splicing, filename)
t_splicing_filename <- str_replace(t_splicing_filename, ".gb", ".fasta")
glimpse(t_splicing_filename)
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refgenes/gmap/triticum/")
write.table(t_splicing_filename, file = "wheat_splicing.csv", append = FALSE, quote = FALSE, sep = " ", dec = ".",
            row.names = FALSE, col.names = FALSE)

# group hordeum by molecule
h_nosplicing <- dplyr::filter(hordeum, molecule %in% c("genomic DNA"))
glimpse(h_nosplicing)
# save filename-list
h_nosplicing_filename <- pull(h_nosplicing, filename)
h_nosplicing_filename <- str_replace(h_nosplicing_filename, ".gb", ".fasta")
glimpse(h_nosplicing_filename)
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refgenes/gmap/hordeum/")
write.table(h_nosplicing_filename, file = "barley_nosplicing.csv", append = FALSE, quote = FALSE, sep = " ", dec = ".",
            row.names = FALSE, col.names = FALSE)

h_splicing <- dplyr::filter(hordeum, molecule %in% c("unassigned DNA", "transcribed RNA", "mRNA"))
glimpse(h_splicing)
# save
h_splicing_filename <- pull(h_splicing, filename)
h_splicing_filename <- str_replace(h_splicing_filename, ".gb", ".fasta")
glimpse(h_splicing_filename)
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refgenes/gmap/hordeum/")
write.table(h_splicing_filename, file = "barley_splicing.csv", append = FALSE, quote = FALSE, sep = " ", dec = ".",
            row.names = FALSE, col.names = FALSE)

# group aegilops by molecule
a_nosplicing <- dplyr::filter(aegilops, molecule %in% c("genomic DNA"))
glimpse(a_nosplicing)
# save filename-list
a_nosplicing_filename <- pull(a_nosplicing, filename)
a_nosplicing_filename <- str_replace(a_nosplicing_filename, ".gb", ".fasta")
glimpse(a_nosplicing_filename)
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refgenes/gmap/aegilops/")
write.table(a_nosplicing_filename, file = "aegilops_nosplicing.csv", append = FALSE, quote = FALSE, sep = " ", dec = ".",
            row.names = FALSE, col.names = FALSE)

a_splicing <- dplyr::filter(aegilops, molecule %in% c("unassigned DNA", "transcribed RNA", "mRNA"))
glimpse(a_splicing)
# save filename-list
a_splicing_filename <- pull(a_splicing, filename)
a_splicing_filename <- str_replace(a_splicing_filename, ".gb", ".fasta")
glimpse(a_splicing_filename)
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refgenes/gmap/aegilops/")
write.table(a_splicing_filename, file = "aegilops_splicing.csv", append = FALSE, quote = FALSE, sep = " ", dec = ".",
            row.names = FALSE, col.names = FALSE)

Above files had to be concatenated into single `.fasta` files, despite trying to solve it in R (proof below) it turned to be better to use bash:
* `cat $(cat wheat_splicing.csv) > wheat_splicing.fasta` -the source table had to be space separated, command has to be executed in the directory of the `.fasta` files

# concatenate triticum nucleotide_fasta files into single file and save into gmap/triticum folder
### nosplicing
library(seqinr)
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refgenes/two_times_filtered/fasta_nucleotide")
t_nosplicing_cat <- foreach(i=t_nosplicing_filename, .packages ="seqinr") %do% read.fasta(file = i,  as.string = TRUE, forceDNAtolower = FALSE, whole.header = TRUE, set.attributes = FALSE)

head(t_nosplicing_cat)

# DOES NOT WORK!!!! - fasta headers are cut off no matter what options are used
# write into file
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refgenes/gmap/triticum")
#triticum_nosplicing <- tempfile(fileext = "fasta")
#tmpf <- tempfile()
write.fasta(sequences = t_nosplicing_cat, names = names(t_nosplicing_cat), file.out = "./triticum_nosplicing1.fasta", open = "w", nbchar = 60, as.string = TRUE, whole.header = TRUE)
test <- read.fasta("./triticum_nosplicing.fasta", set.attributes = TRUE)
identical(test, t_nosplicing_cat)

## 4.3 Mapping with GMAP
* **official website:** http://research-pub.gene.com/gmap/
* my gmap is installed into `seqtools` conda environment
* due to the large genome size in barley&wheat I ahd to use `gmapl` instead, but it was installed along with gmap
* **perl scripts to inspect files:** 
    - `perl -ne 'print $1,"\n" if /mol_type="([^"]+)"/;' genbank/*.gb| sort | uniq -c`
    - `perl -ne 'print $1,"\n" if /^\s*ORGANISM\s+(.+)/;' genbank/*.gb| sort | uniq -c`
* **preparations:** 
    - gene-seqs belong to 3 organisms: triticum, hordeum, aegilops
    - they have 4 types of data: 
        - genomic DNA - `--nosplicing` flag has to be on! (Turns off splicing (useful for aligning genomic sequences onto a genome)
        - mRNA & transcribedRNA & unassigned (which we assign to transcribedRNA) - `--nosplicing` flag has to be off!
    - create dataframe with "Organism" and "mol_type" information to parse files
    - create concatenated merged fasta file from all the nucleotide gene fastas for each group to map (script and output paths are above)
    - create symlinks of each ref.genome into gmap folder for indexing
* **1.:** indexing
    - directory: `/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/indexes/gmap`
    - run indexing command in a screen session in the directory of the respective genome where symlinks are:
        
        `gmap_build -D /nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/indexes/gmap -d triticum -k 15 genome.fasta`
        
        `gmap_build -D /nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/indexes/gmap -d hordeum -k 15 genome.fasta`
        
        `gmap_build -D /nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/indexes/gmap -d aegilops -k 15 genome.fasta`
    - use the `-D` flag to specify the default location for the tool
    - it can take upto a day long
* **2.:** use Aegilops 2+1 fastas to create 12 different output formats
    - inspect result files and choose the option for the entire analysis
    - for splicing category only a single file was tested
    - for nosplicing the cat csv based 2 files
* **3.:** run the respective `gmap_gff.sh` script on slurm
    - here `-f` specifies the output file format, where `est.gff` and `alignment.gff` is the most recommended and can be later converted into counttable using STAR/HISAT
    - destination directories: `/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refgenes/gmap/`
    - scripts are found: `home/pgsb/vanda.marosi/scripts/gmap/`

## Location of Ref.Genomes
* all of them are symlinked under: `/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/genomes/`

`./Horvu/CDS.fasta:49281
./Horvu/protein.fasta:46294
./Horvu/transcript.fasta:49281
./Horvu/genome.fasta:8
./Triae/protein.fasta:122722
./Triae/genome.fasta:22
./Triae/transcript.fasta:123075
./Triae/CDS.fasta:122722
./Aegta/transcript.fasta:68789
./Aegta/genome.fasta:8
./Aegta/CDS.fasta:68789
./Aegta/protein.fasta:68390`

## 4.4 Filtering with identity&coverage values and intersection with ref.annotations

sessionInfo()
