==> 04_Ref_samples_download_preprocess.ipynb <==
# Creating datasets: Download & Symlink & Preprocessing steps
## 1. Download (missing = not yet downloaded) data

library(tidyverse, warn.conflicts = FALSE)

setwd("/home/vanda.marosi//floral_development_thesis_vm/datatables/")
# import my datasets
wheat <- read.table("wheat_final.csv", header = TRUE, sep = ",", stringsAsFactors = FALSE)
barley <- read.table("barley_final.csv", header = TRUE, sep = ",", stringsAsFactors = FALSE)
glimpse(wheat)
glimpse(barley)
# import Nadia & Maxim's datasets
wheat_m <- read.table("wheat_22012020_v1_MM.csv", header = TRUE, sep = ",", stringsAsFactors = FALSE)
barley_n <- read.table("hordeum_vulgare_14012020_NK.csv", header = TRUE, sep = ",", stringsAsFactors = FALSE)
glimpse(wheat_m)
glimpse(barley_n)

# intersect wheat tables
w_v <- select(wheat, Run.ID, Dataset.name, NCBI.BioProject)
w_m <- select(wheat_m, sample)
colnames(w_m) <- "Run.ID"
# this gives id-s that appear in mine but not in Maxim's -> 67 fastq-s have to be downloaded
y <- anti_join(w_v, w_m, by = "Run.ID") # all rows in w_v that do not have a match in w_m
glimpse(y)

# this gives id-s that appear in Maxim's from my list -> 148 fastq-s have to be symbolic linked
y1 <- inner_join(w_v, w_m, by = "Run.ID") # retain only rows in both sets
glimpse(y1)

# intersect barley tables
b_v <- select(barley, Run.ID, Dataset.name, NCBI.BioProject)
b_n <- select(barley_n, Run)
colnames(b_n) <- "Run.ID"
# this gives id-s that appear in mine but not in Nadia's -> 65 fastq-s have to be downloaded
x <- anti_join(b_v, b_n, by = "Run.ID")
glimpse(x)

# this gives id-s that appear in Maxim's from my list -> 175 fastq-s have to be symbolic linked
x1 <- inner_join(b_v, b_n, by = "Run.ID")
glimpse(x1)

## 1.1 Download additional fastq-s (from my dataset) using `sra-toolkit`

# get Run.ID-s into single vector
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/triticum")
w_download <- dplyr::pull(y, Run.ID)
write.table(w_download, file = "wheat_download_runids.txt", append = FALSE, quote = FALSE, sep = "\t", dec = ".",
            row.names = FALSE, col.names = FALSE)
glimpse(w_download)
    
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/hordeum")
b_download <- dplyr::pull(x, Run.ID)
write.table(b_download, file = "barley_download_runids.txt", append = FALSE, quote = FALSE, sep = "\t", dec = ".",
            row.names = FALSE, col.names = FALSE)
glimpse(b_download)

## 1.2 Steps to download extra files (added to the flower-dataset but not listed in Nadia&Maxim's tables)
1. N&M's and my selected samples Run.ID-s were intersected and saved as vectors: 67 wheat, 65 barley samples needed to be downloaded
2. The above vectors were written into .txt files, placed in subfolders for the respective species
3. SRA Toolkit was downloaded to the same main folder (https://www.ncbi.nlm.nih.gov/sra/docs/sradownload/)
4. `PATH-bin` had to be established every time I opened the terminal, solved this issue by placing it in my bashrc
5. `vdb-config -i` was called for basic configuration following the website's recommendation, and a main empty folder was put as download-target
6. `prefetch --option-file wheat_download_runids.txt`
7. `fasterq-dump --split-files *.sra`
8. copied the `.*fastq`-s into destination folder (triticum), and emptied the sra-download folder for the next batch
9. `prefetch --option-file barley_download_runids.txt`
10. `fasterq-dump --split-files *.sra`
11. copied the `.*fastq`-s into destination folder (hordeum), and emptied the sra-download folder
12. for the extra 2 China + 1 France dataset, the respective websites were called, each of the fastq-s weblink were copied into a bash script with `wget` 
13. `chmod +x` - to make the bash-scripts executable
14. `./*.sh` - to execute them
15. I double-checked whether all files were downloaded, manually downloaded the missing ones, eg.: `prefetch SRR000001` and than `fasterq-dump --split-files SRR000001.sra`

## 2. Create symlinks 

# get Run.ID-s into single vector
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/triticum")
#w_download <- as.vector(dplyr::pull(y, Run.ID))
w_symlink <- dplyr::pull(y1, Run.ID)
#w_download <- as.name(w_download)
write.table(w_symlink, file = "wheat_symlink_runids.txt", append = FALSE, quote = FALSE, sep = ",", dec = ".",
            row.names = FALSE, col.names = FALSE)
glimpse(w_symlink)
    
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/hordeum")
b_symlink <- dplyr::pull(x1, Run.ID)
write.table(b_symlink, file = "barley_symlink_runids.txt", append = FALSE, quote = FALSE, sep = ",", dec = ".",
            row.names = FALSE, col.names = FALSE)
glimpse(b_symlink)

## 2.1 My steps to create symbolic-links
1. N&M's and my selected samples Run.ID-s were intersected and saved as vectors: 148 wheat, 175 barley samples needed to be symbolic-linked
2. The above vectors were written into `wheat/barley_symlink_runids.txt` files, placed in subfolders for the respective species

### For Wheat dataset:
* from `/nfs/scratch/daniel.lang/comparative_triticeae/raw/T.aestivum`
* to `/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/triticum`

### For Barley dataset:
* from `/nfs/scratch/daniel.lang/comparative_triticeae/raw/H.vulgare`
* to `/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/hordeum`

## 2.2 Bash-scripts to find failed SRA-s and create symlinks for existing ones:
1. `csver.sh`: creates a `.csv` file with two columns, 1st column: source files with their path, 2nd column: final destination with symlink-names
    - it is also optimized to detect first the paired fastq-s, and as second step also the single ones (for later use of creating input tables to FastQC/Trimmomatic etc)
2. `confirmationlinker.sh`: 
    - #1 run: which files from my link-list has missing targets from source directory? points them out, that they can be manually downloaded
    - #2 run: creates link between two tables to prepare them for `ln -s`
3. `linker.sh`: creates symlinks using `ln -s /path/to/file /path/to/link`

* scripts are located at: `vanda.marosi/scripts/`

* for the wheat table missing SRA-s(failed downloads), I have `prefetch`-ed them into my folder: SRR7106399, SRR7106400, SRR7106402, SRR7106404, SRR6802615

* for the barley table missing SRA-s are 87, detected via the `confirmationlinker.sh` and cut into new `missing_barley.txt` and `prefetch`-ed:

    from this (column of 87 rows):
    **this link will have no target ERR2026425.fastq.gz**
    
    made this (column of 87 rows):
    **ERR2026425**

    `./confirmationlinker.sh | cut -d " " -f 7 | cut -d "." -f 1 > missing_barley.txt`
* to set maximum file size download limit to 100G (default: 20G): `prefetch --max-size 100G SRR649944`


## 2.3 Create symlinks for French dataset
* used the three above mentioned bash scripts, modified partially but creating symlinks into same folder, with these names: FRR7482-9

setwd("~/scripts")
symlink_french_dataset <- read.table("ngtmpindex.csv", header = FALSE, sep = ",", stringsAsFactors = FALSE)
head(symlink_french_dataset)

## 3. Crosscheck downloaded files with original table

# triticum double-check, here is a column created from the downloaded files using bash (ls *gz | cut -d "." -f 1 | cut -d "_" -f 1 | uniq | grep -v NG > wheat_final_list.txt)
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/triticum")
w_check <- read.table("wheat_final_list.txt", header = FALSE, sep = "\t", stringsAsFactors = FALSE)
glimpse(w_check)
colnames(w_check) <- "Run.ID"
# the intersection with my original wheat table (based on which files were downloaded) will give the rows that are present in both tables
ww <- inner_join(w_v, w_check, by = "Run.ID")
glimpse(ww)
# everything was succesfull!

# hordeum double-check -> here is a column created from the downloaded files using bash (ls *gz | cut -d "." -f 1 | cut -d "_" -f 1 | uniq | grep -v NG > barley_final_list.txt
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/hordeum")
b_check <- read.table("barley_final_list.txt", header = FALSE, sep = "\t", stringsAsFactors = FALSE)
glimpse(b_check)
colnames(b_check) <- "Run.ID"
# the intersection with my original wheat table (based on which files were downloaded) will give the rows that are present in both tables
bb <- inner_join(b_v, b_check, by = "Run.ID")
glimpse(bb)
# everything was succesfull!

## 4. Rename libraries
* to create  universal `_1/2_fastq.gz` naming using bash scripts
* 1. format in triticum only: `CRR088946_f1.fastq.gz`
    solution: `for file in $(find ./*f1.fastq.gz);do file1=$(echo $file | cut -d '/' -f 2 | cut -d '_' -f 1 ); echo mv ./${file1}_f1.* ./${file1}1.fastq.gz; done`
* 2. format in triticum only: `CRR088946_r2.fastq.gz`
    solution: `for file in $(find ./*r2.fastq.gz);do file1=$(echo $file | cut -d '/' -f 2 | cut -d '_' -f 1 ); echo mv ./${file1}_r2.* ./${file1}_2.fastq.gz; done`
* 3. format in tri/hordeum both: `SRR7106404.sra_1.fastq.gz`
    solution: `for file in $(find ./*.sra_1.fastq.gz);do file1=$(echo $file | cut -d '/' -f 2 | cut -d '.' -f 1 ); echo mv ./${file1}.sra_1* ./${file1}_1.fastq.gz; done`
* 4. format in tri/hordeum both: `SRR7106404.sra_2.fastq.gz`
    solution: `for file in $(find ./*.sra_2.fastq.gz);do file1=$(echo $file | cut -d '/' -f 2 | cut -d '.' -f 1 ); echo mv ./${file1}.sra_2* ./${file1}_2.fastq.gz; done`
* 5. format in tri/hordeum both: `SRR5464507.sra.fastq.gz`
     solution: `for file in $(find ./*.sra.fastq.gz);do file1=$(echo $file | cut -d '/' -f 2 | cut -d '.' -f 1 ); echo mv ./${file1}.sra.* ./${file1}.fastq.gz; done`

## 5. MD5 checksum
* Perl script from Daniel: 
    - for triticum: `perl -ne 'chomp; @a=split/\t/; next if $a[0]=~ /^CRR|FRR/; my $u= sprintf(q(https://www.ebi.ac.uk/ena/data/warehouse/filereport?accession=%s&result=read_run&fields=run_accession,fastq_ftp,fastq_md5,fastq_bytes,read_count,first_public,last_updated,center_name,broker_name),$a[0]); my $c=`curl -s "$u"`;  if ($i) {$c=~ s/^run_accession.+\n//g;} print $c; $i++;' wheat_project_table.txt > wheat_project_table.only_sra.metadata.txt`
    - for hordeum: `perl -ne 'chomp; @a=split/\t/; next if $a[0]=~ /^CRR|FRR/; my $u= sprintf(q(https://www.ebi.ac.uk/ena/data/warehouse/filereport?accession=%s&result=read_run&fields=run_accession,fastq_ftp,fastq_md5,fastq_bytes,read_count,first_public,last_updated,center_name,broker_name),$a[0]); my $c=`curl -s "$u"`;  if ($i) {$c=~ s/^run_accession.+\n//g;} print $c; $i++;' barley_project_table.txt > barley_project_table.only_sra.metadata.txt`

* 3 datasets that were downloaded from other websites had to be excluded as they dont have SRAnumber: `CRR|FRR`:
    - CRR: 1x18, 1x27
    - FRR: 1x8
    - alltogether 53 samples are impossible to md5sum-check

### 5.1 Triticum

setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/triticum")
md5_myfiles_t <- read.table("md5sum_triticum.txt", header = FALSE, stringsAsFactors = FALSE, fill = TRUE)
colnames(md5_myfiles_t) <- c("md5sum", "read_origin")
glimpse(md5_myfiles_t)
# in my files there are total 395 files: 180 paired (2x=360) + 35 single reads
tail(md5_myfiles_t)

md5_srafiles_t <- read.table("wheat_project_table.only_sra.metadata.txt", header = TRUE, stringsAsFactors = FALSE, fill = TRUE)
glimpse(md5_srafiles_t)
# from my files 53 had to be excluded (not SRA-based), and 8 are single reads, and 9 SRAquery failed = (395-53-8-9)/2=162 rows
# so 9 rows are fully empty!!! only run accession is available: rows 96-104 are empty
tail(md5_srafiles_t)

# wrangling md5_srafiles_t
# delete missing gap of 94-104
md5_1 <- slice(md5_srafiles_t, 1:95)
md5_2 <- slice(md5_srafiles_t, 105:162)
md5_srafiles_joined <- bind_rows(md5_1, md5_2)
# select relevant columns
md5_srafiles_js <- select(md5_srafiles_joined, run_accession, fastq_ftp, fastq_md5)
# separate read names
md5_srafiles_t_n <- separate(md5_srafiles_js, fastq_ftp, sep=";", into=c("read", "read2"), remove = TRUE)
# separate md5sums
md5_srafiles_t_nm <- separate(md5_srafiles_t_n, fastq_md5, sep=";", into=c("md5sum", "md5sum2"), remove = TRUE)
glimpse(md5_srafiles_t_nm)
# bind column2-s under column1-s as new rows
col1 <- select(md5_srafiles_t_nm, read, md5sum)
col2 <- select(md5_srafiles_t_nm, read2, md5sum2)
colnames(col2) <- c("read", "md5sum")
md5_srafiles_clean <- bind_rows(col1, col2)
glimpse(md5_srafiles_clean)

# delete missing gap of 241-248 from single reads
md5_j1 <- slice(md5_srafiles_clean, 1:240)
md5_j2 <- slice(md5_srafiles_clean, 249:306)
md5_srafiles_jc <- bind_rows(md5_j1, md5_j2)
# final table
glimpse(md5_srafiles_jc)
head(md5_srafiles_jc)

# intersect two tables
missing <- anti_join(md5_srafiles_jc, md5_myfiles_t, by = "md5sum")
glimpse(missing)

### 5.2 Hordeum

# import tables
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/hordeum")
md5_myfiles_h <- read.table("md5sum_hordeum.txt", header = FALSE, stringsAsFactors = FALSE, fill = TRUE)
colnames(md5_myfiles_h) <- c("md5sum", "read_origin")
glimpse(md5_myfiles_h)
# in my files there are total 433 files: 193 paired (2x=386) + 47 single reads
tail(md5_myfiles_h)

md5_srafiles_h <- read.table("barley_project_table.only_sra.metadata.txt", header = FALSE, stringsAsFactors = FALSE, fill = TRUE)
colnames(md5_srafiles_h) <- c("run_accession", "fastq_ftp", "fastq_md5", "4", "5", "6", "7", "8", "9", "10", "11")
md5_srafiles_h <- select(md5_srafiles_h, run_accession, fastq_ftp, fastq_md5)
glimpse(md5_srafiles_h)
# there are many empty queries!!!
tail(md5_srafiles_h)

# wrangling md5_srafiles_h
# separate read names
md5_srafiles_t_n <- separate(md5_srafiles_h, fastq_ftp, sep=";", into=c("read", "read2"), remove = TRUE)
# separate md5sums
md5_srafiles_t_nm <- separate(md5_srafiles_t_n, fastq_md5, sep=";", into=c("md5sum", "md5sum2"), remove = TRUE)
glimpse(md5_srafiles_t_nm)
# bind column2-s under column1-s as new rows
col1 <- select(md5_srafiles_t_nm, read, md5sum)
col2 <- select(md5_srafiles_t_nm, read2, md5sum2)
colnames(col2) <- c("read", "md5sum")
md5_srafiles_clean <- bind_rows(col1, col2)
glimpse(md5_srafiles_clean)

# delete missing gap of 241-248 from single reads
md5_h1 <- slice(md5_srafiles_clean, 1:268)
md5_h2 <- slice(md5_srafiles_clean, 316:536)
md5_srafiles_final <- bind_rows(md5_h1, md5_h2)
# final table
glimpse(md5_srafiles_final)
head(md5_srafiles_final)

# intersect two tables
missing <- anti_join(md5_srafiles_final, md5_myfiles_h, by = "md5sum")
glimpse(missing)

# 6. Preprocessing
* 1. FastQC + MultiQC
* 2. Trimmomatic
* 3. FastQC + MultiQC

## 6.1 Create Project table for FastQC_01_raw
### 6.1.1 Triticum

w_pt <- select(wheat, Run.ID, Dataset.name, Tissue)
w_pt <- inner_join(w_pt, w_check, by = "Run.ID")
colnames(w_pt) <- c("ID", "dataset_name", "tissue", "library")
glimpse(w_pt)
# save project table
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/triticum")
write.table(w_pt, file = "wheat_project_table.txt", append = FALSE, quote = FALSE, sep = "\t", dec = ".",
            row.names = FALSE, col.names = FALSE)
# create & save testing subset 
w_pt_test1 <- slice(w_pt, 1:2) # choosing 2 paried reads with "_1/2.fastq.gz"
w_pt_test2 <- slice(w_pt, 102:103) # choosing 2 single reads with "fq.gz"
w_pt_test3 <- slice(w_pt, 141) # choosing 1 single read with "sra.fastq.gz" extension
w_pt_test <- bind_rows(w_pt_test1, w_pt_test2)
w_pt_test <- bind_rows(w_pt_test, w_pt_test3)
print(w_pt_test)
write.table(w_pt_test, file = "wheat_project_table_test.txt", append = FALSE, quote = FALSE, sep = "\t", dec = ".",
            row.names = FALSE, col.names = TRUE)

### 6.1.2 Hordeum

b_pt <- select(barley, Run.ID, Dataset.name, Tissue)
b_pt <- inner_join(b_pt, b_check, by = "Run.ID")
colnames(b_pt) <- c("ID", "dataset_name", "tissue")
glimpse(b_pt)
# save project table
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/hordeum")
write.table(b_pt, file = "barley_project_table.txt", append = FALSE, quote = FALSE, sep = "\t", dec = ".",
            row.names = FALSE, col.names = TRUE)


### 6.1.3 Delete 0 size files from FastQC-folder
1. to confirm: `for file in $(find ./ -size 0 -print); do ls $file; echo $file will be removed;ls $file -lah; done`

2. to delete: `for file in $(find ./ -size 0 -print); do ls $file; echo $file will be removed;ls $file -lah; rm $file; done`

### 6.1.4 Crosscheck fastqc-s with original dataset

# hordeum had all the runs completed
# triticum had one paired read missing:
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/triticum/01_FastQC_raw_paired")
final <- read.table("01_trit_paired_fastqc.txt", header = FALSE, sep = "\t", stringsAsFactors = FALSE)
colnames(final) <- "Run.ID"
glimpse(final)

# create table with read layout
w_check <- select(wheat, Run.ID, Library.layout)
w_check_paired <- group_by(w_check, Library.layout)
glimpse(w_check_paired)

# intersect tables and get the missing paired read
joined_wf <- anti_join(w_check_paired, final, by = "Run.ID")
glimpse(joined_wf)
print(joined_wf)
# for some reason SRR10737427* wasnt processed, although it exists as a symlink and looks whole

# create extra project table for that 1 missing file: SRR10737427
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/triticum")
extra <- read.table("wheat_project_table_trimmomatic_paired.txt", header = FALSE, sep = "\t", stringsAsFactors = FALSE)
extra <- slice(extra, 2)
colnames(extra) <- c("ID", "dataset_name", "tissue")
glimpse(extra)
write.table(extra, file = "extra_1missingfile_paired.txt", append = FALSE, quote = FALSE, sep = "\t", dec = ".",
            row.names = FALSE, col.names = TRUE)

## 6.2 Trimming with `Trimmomatic`
* bash scripts are available: `home/pgsb/vanda.marosi/scripts/triticum & /hordeum`

### Create separate project tables for Trimmomatic paired & single reads
### 6.2.1 Triticum

setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/triticum")
w_pt_trim <- select(wheat, Run.ID, Dataset.name, Tissue, Library.layout)
colnames(w_pt_trim) <- c("ID", "dataset_name", "tissue", "library")
w_pt_trim <- group_by(w_pt_trim, library)
glimpse(w_pt_trim)

w_pt_trim_paired <- filter(w_pt_trim, library == "paired")
w_pt_trim_paired <- ungroup(w_pt_trim_paired)
w_pt_trim_paired <- select(w_pt_trim_paired, ID, dataset_name, tissue)
glimpse(w_pt_trim_paired)
write.table(w_pt_trim_paired, file = "wheat_project_table_trimmomatic_paired.txt", append = FALSE, quote = FALSE, sep = "\t", dec = ".",
            row.names = FALSE, col.names = TRUE)

w_pt_trim_single <- filter(w_pt_trim, library == "single")
w_pt_trim_single <- ungroup(w_pt_trim_single)
w_pt_trim_single <- select(w_pt_trim_single, ID, dataset_name, tissue)
glimpse(w_pt_trim_single)
write.table(w_pt_trim_single, file = "wheat_project_table_trimmomatic_single.txt", append = FALSE, quote = FALSE, sep = "\t", dec = ".",
            row.names = FALSE, col.names = TRUE)

### 6.2.2 Hordeum

setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/hordeum")
b_pt_trim <- select(barley, Run.ID, Dataset.name, Tissue, Library.layout)
colnames(b_pt_trim) <- c("ID", "dataset_name", "tissue", "library")
b_pt_trim <- group_by(b_pt_trim, library)
glimpse(b_pt_trim)

b_pt_trim_paired <- filter(b_pt_trim, library == "paired")
b_pt_trim_paired <- ungroup(b_pt_trim_paired)
b_pt_trim_paired <- select(b_pt_trim_paired, ID, dataset_name, tissue)
glimpse(b_pt_trim_paired)
write.table(b_pt_trim_paired, file = "barley_project_table_trimmomatic_paired.txt", append = FALSE, quote = FALSE, sep = "\t", dec = ".",
            row.names = FALSE, col.names = TRUE)

b_pt_trim_single <- filter(b_pt_trim, library == "single")
b_pt_trim_single <- ungroup(b_pt_trim_single)
b_pt_trim_single <- select(b_pt_trim_single, ID, dataset_name, tissue)
glimpse(b_pt_trim_single)
write.table(b_pt_trim_single, file = "barley_project_table_trimmomatic_single.txt", append = FALSE, quote = FALSE, sep = "\t", dec = ".",
            row.names = FALSE, col.names = TRUE)

## 6.3 FastQC & MultiQC
* bash scripts are available: `home/pgsb/vanda.marosi/scripts/triticum & /hordeum`

### Results
* **wheat paired:** 1 sample failed - out of 180
* **wheat single:** all 35 good
* **barley single:** all 47 good
* **barley paired:** 1 sample failed - out of 
 - since then all of them checked and are good!

#### Investigating 1 barley paired failed sequences
1. table of successful runs with bash: `ls | grep ".zip" | cut -d "_" -f 1 | uniq > 03_hordeum_paired_fastqc.txt` 

## looking for barley paired 2 missing files
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/hordeum/03_FastQC_trimmed_paired")
fqc_b <- read.table("03_hordeum_paired_fastqc.txt", header = FALSE, sep = "\t", stringsAsFactors = FALSE)
colnames(fqc_b) <- "Run.ID"
glimpse(fqc_b)

# create table with read layout
b_check <- select(barley, Run.ID, Library.layout)
b_check_paired <- filter(b_check, Library.layout == "paired")
glimpse(b_check_paired)

# intersect tables and get the missing paired read
joined_b <- anti_join(b_check_paired, fqc_b, by = "Run.ID")
print(joined_b)

read ERR1248084 has 101 basepair, and seems to be an average quality sample from the 1st (raw) multiqc
* solution: rerun sample from trimmomatic separately, bash scripts are made with name `extra` for trimming and last fastqc

# create extra project table for that 1 missing file: ERR1248084
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/hordeum")
extra <- read.table("barley_project_table_trimmomatic_paired.txt", header = FALSE, sep = "\t", stringsAsFactors = FALSE)
extra <- slice(extra, 2)
colnames(extra) <- c("ID", "dataset_name", "tissue")
glimpse(extra)
write.table(extra, file = "extra_1missingfile_paired.txt", append = FALSE, quote = FALSE, sep = "\t", dec = ".",
            row.names = FALSE, col.names = TRUE)

after re-running:
- triticum SRR10737427 paired sample: 01_FastQC & `multiqc -f .`, Trimmomatic, 03_FastQC & `multiqc -f .`
- hordeum ERR1248084 paired sample: Trimmomatic, 03_FastQC & `multiqc -f .`, 

a new multiqc-report was generated and finally all samples were fully processed and in place

# Final results of quality controlled data
* created bash scripts to measure data size in GB
* scripts are available under `size_calculator_wheat/barley.sh` in the Bash_scripts folder
* final calculation: 
       barley = 696 GB paried + 38 GB single = 734 GB
       wheat = 723 GB paired + 19 GB single = 742 GB
       in total 1476 GB = 1.44 TB data

sessionInfo()
