==> 05_Ref_samples_mapping.ipynb <==
# Mapping 

## 1. Mapping single & paired reads with `kallisto` using bash scripts on slurm
* adapted Maxim's scripts:
* created symlinks: ` ln -s /path/to/file /path/to/symlink`
* directory of indexes: `/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/indexes/kallisto`
* created indexes using: - with a k-mer length of 31
    - triticum: `kallisto index -i /nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/indexes/kallisto/triticum_transcritpt.idx transcript.fasta`
    - hordeum: `kallisto index -i /nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/indexes/kallisto/hordeum_transcript.idx transcript.fasta`
* location of kallisto tool: `/home/pgsb/vanda.marosi/anaconda3/envs/seqtools/bin/kallisto`
* bash scripts are available: `~/scripts/triticum OR hordeum/run_04_..._kallisto_single OR paired.sh`

## 1.1 Location of Ref.Genomes
* all of them are symlinked under: `/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/genomes/`
`./Horvu/CDS.fasta:49281
./Horvu/protein.fasta:46294
./Horvu/transcript.fasta:49281
./Horvu/genome.fasta:8
./Triae/protein.fasta:122722
./Triae/genome.fasta:22
./Triae/transcript.fasta:123075
./Triae/CDS.fasta:122722`

* as an example Daniel's snakemake kallisto pipeline: https://ibis-gitlab.helmholtz-muenchen.de/daniel.lang/kallisto/-/tree/master

## 1.2 After mapping
* the default result from `kallisto` for each sample is 3 files: 
    - `abundance.h5` - binary file
    - `abundance.tsv` - tab separated file: `target_id, length, eff_length, est_counts, tpm`
    - `run_info.json` - inf about run command, time, tool version etc.
    - kallisto created a normalized table of the transcripts, which can be used for PCA already
    - but to divide the transcripts into real genes `readr` and `tximport` packages has to be used to combine my results with annotation.gff files and make a real count-table
    - count table can be log-transformed, and further statistical validations, PCAs using `r-sleuth`, `DeSeq2`, `HR`
* manuals & example guidelines:
    - tximport: https://bioconductor.org/packages/release/bioc/vignettes/tximport/inst/doc/tximport.html#kallisto

# 2. Map transcripts to genes and create count-tables
* here first only paired OR single reads were converted into a count-table and later was their merged table made

library(rhdf5, warn.conflicts = FALSE)
library(readr, warn.conflicts = FALSE)
library(tximport, warn.conflicts = FALSE)
library(GenomicFeatures, warn.conflicts = FALSE)
# do not call tidyverse here!!! its `select()` interferes with tximport!
# call tidyverse only after creating all necessary tx2gene objects

## Triticum paired

setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/triticum/")
trit_paired <- read.table("wheat_project_table_trimmomatic_paired.txt", sep = "\t", header = TRUE, stringsAsFactors = FALSE)
str(trit_paired)

setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/triticum/04_Kallisto_paired")
files_tp <- file.path(trit_paired$ID, "abundance.h5")
names(files_tp) <- paste0(trit_paired$ID)
head(files_tp)
all(file.exists(files_tp))

Transcripts need to be associated with gene IDs for gene-level summarization:
1. We first make a data.frame called tx2gene with two columns: 
    - 1) transcript ID and 
    - 2) gene ID. 
2. The column names do not matter but this column order must be used. The transcript ID must be the same one used in the abundance files.

# create tx2gene data.frame
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/genomes/Triae/")
txdb_trit <- makeTxDbFromGFF("annotation.gff3", organism = "Triticum aestivum")
#makeTxDbFromGFF(file, format=c("auto", "gff3", "gtf"), dataSource=NA, organism=NA, taxonomyId=NA, circ_seqs=DEFAULT_CIRC_SEQS, chrominfo=NULL, miRBaseBuild=NA, dbxrefTag)
str(txdb_trit)

k_trit <- keys(txdb_trit, keytype = "TXNAME")
tx2gene_trit <- select(txdb_trit, k_trit, "GENEID", "TXNAME")
head(tx2gene_trit)

# save gene-mapping inf for r-sleuth analysis later
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/triticum/")
write.table(tx2gene_trit, file = "wheat_tx2gene.tsv", append = FALSE, quote = FALSE, sep = "\t", dec = ".",
            row.names = TRUE, col.names = TRUE)

transcripts can be converted into count-table without assigning them to genes: 
* `txi = tximport(files, type = "kallisto", txOut = TRUE)`
* `head(txi$counts)`

# create count table
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/triticum/04_Kallisto_paired/")
txi_trit_paired <- tximport(files_tp, type = "kallisto", tx2gene = tx2gene_trit)
head(txi_trit_paired$counts)

# save count table
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/triticum/")
write.table(txi_trit_paired, file = "wheat_count_table_paired.txt", append = FALSE, quote = FALSE, sep = "\t", dec = ".",
            row.names = TRUE, col.names = TRUE)
saveRDS(txi_trit_paired, file = "wheat_count_table_paired.rds")

### Triticum single

setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/triticum/")
trit_single <- read.table("wheat_project_table_trimmomatic_single.txt", sep = "\t", header = TRUE, stringsAsFactors = FALSE)
glimpse(trit_single)

setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/triticum/04_Kallisto_single")
files_ts <- file.path(trit_single$ID, "abundance.h5")
names(files_ts) <- paste0(trit_single$ID)
head(files_ts)
all(file.exists(files_ts))

# create count table
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/triticum/04_Kallisto_single/")
txi_trit_single <- tximport(files_ts, type = "kallisto", tx2gene = tx2gene_trit)
head(txi_trit_single$counts)

# save count table
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/triticum/")
write.table(txi_trit_single, file = "wheat_count_table_single.txt", append = FALSE, quote = FALSE, sep = "\t", dec = ".",
            row.names = TRUE, col.names = TRUE)
saveRDS(txi_trit_single, file = "wheat_count_table_single.rds")

### Hordeum paired

setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/hordeum/")
hord_paired <- read.table("barley_project_table_trimmomatic_paired.txt", sep = "\t", header = TRUE, stringsAsFactors = FALSE)
glimpse(hord_paired)

setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/hordeum/04_Kallisto_paired")
files_hp <- file.path(hord_paired$ID, "abundance.h5")
names(files_hp) <- paste0(hord_paired$ID)
head(files_hp)
all(file.exists(files_hp))

# create tx2gene data.frame
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/genomes/Horvu/")
txdb_hord <- makeTxDbFromGFF("annotation.gff3", organism = "Hordeum vulgare")
#makeTxDbFromGFF(file, format=c("auto", "gff3", "gtf"), dataSource=NA, organism=NA, taxonomyId=NA, circ_seqs=DEFAULT_CIRC_SEQS, chrominfo=NULL, miRBaseBuild=NA, dbxrefTag)
str(txdb_hord)

k_hord <- keys(txdb_hord, keytype = "TXNAME")
tx2gene_hord <- select(txdb_hord, k_hord, "GENEID", "TXNAME")
head(tx2gene_hord)

# save gene-mapping inf for r-sleuth analysis later
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/hordeum/")
write.table(tx2gene_hord, file = "barley_tx2gene.tsv", append = FALSE, quote = FALSE, sep = "\t", dec = ".",
            row.names = TRUE, col.names = TRUE)

# create count table
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/hordeum/04_Kallisto_paired/")
txi_hord_paired <- tximport(files_hp, type = "kallisto", tx2gene = tx2gene_hord)
head(txi_hord_paired$counts)

# save count table
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/hordeum/")
write.table(txi_hord_paired, file = "barley_count_table_paired.txt", append = FALSE, quote = FALSE, sep = "\t", dec = ".",
            row.names = FALSE, col.names = TRUE)
saveRDS(txi_hord_paired, file = "barley_count_table_paired.rds")

### Hordeum single

setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/hordeum/")
hord_single <- read.table("barley_project_table_trimmomatic_single.txt", sep = "\t", header = TRUE, stringsAsFactors = FALSE)
glimpse(hord_single)

setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/hordeum/04_Kallisto_single")
files_hs <- file.path(hord_single$ID, "abundance.h5")
names(files_hs) <- paste0(hord_single$ID)
head(files_hs)
all(file.exists(files_hs))

# create count table
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/hordeum/04_Kallisto_single/")
txi_hord_single <- tximport(files_hs, type = "kallisto", tx2gene = tx2gene_hord)
head(txi_hord_single$counts)

# save count table
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/hordeum/")
write.table(txi_hord_single, file = "barley_count_table_single.txt", append = FALSE, quote = FALSE, sep = "\t", dec = ".",
            row.names = FALSE, col.names = TRUE)
saveRDS(txi_hord_single, file = "barley_count_table_single.rds")

# Merging paired & single reads
#### 1, symlink single & paried reads into a merge folder using `ln -s /path/to/file /path/to/link`:
    * for barley
        - `ln -s /nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/hordeum/04_Kallisto_single/* /nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/hordeum/04_Kallisto_allreads_symlinked/`
        - `ln -s /nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/hordeum/04_Kallisto_paired/* /nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/hordeum/04_Kallisto_allreads_symlinked/`
    * for wheat
        - `ln -s /nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/triticum/04_Kallisto_single/* /nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/triticum/04_Kallisto_allreads_symlinked/`
        - `ln -s /nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/triticum/04_Kallisto_paired/* /nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/triticum/04_Kallisto_allreads_symlinked/`
#### 2, create file path with all the reads from this folder
### Hordeum

# load tidyverse only after created tx2gene objects, because commands interfere with each-other!!
library(tidyverse, warn.conflicts = FALSE)
setwd("/home/vanda.marosi/floral_development_thesis_vm/datatables/")
barley_meta <- read.table("barley_final.csv", header = TRUE, sep = ",", stringsAsFactors = FALSE)
barley_meta <- select(barley_meta, Run.ID, Dataset)
# #exclude outliers: ERR781040, ERR1457187
barley_meta <- filter(barley_meta, Run.ID != "ERR781040")
barley_meta <- filter(barley_meta, Run.ID != "ERR1457187")
glimpse(barley_meta)
colnames(barley_meta) <- c("ID", "dataset")

setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/hordeum/04_Kallisto_allreads_symlinked/")
files_h <- file.path(barley_meta$ID, "abundance.h5")
names(files_h) <- paste0(barley_meta$ID)
head(files_h)
all(file.exists(files_h))

# use tx2gene data.frame made earlier as txdb_hord and tx2gene_hord
# create count table
txi_hord <- tximport(files_h, type = "kallisto", tx2gene = tx2gene_hord)
dim(txi_hord$counts)
head(txi_hord$counts)

# save count table
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/hordeum/")
write.table(txi_hord, file = "barley_count_table_merged_nooutlier.txt", append = FALSE, quote = FALSE, sep = "\t", dec = ".",
            row.names = FALSE, col.names = TRUE)
saveRDS(txi_hord, file = "barley_count_table_merged_nooutlier.rds")

summary(txi_hord)

### Triticum

setwd("/home/vanda.marosi/floral_development_thesis_vm/datatables/")
wheat_meta <- read.table("wheat_final.csv", header = TRUE, sep = ",", stringsAsFactors = FALSE)
wheat_meta <- select(wheat_meta, Run.ID, Dataset, )
colnames(wheat_meta) <- c("ID", "dataset")
glimpse(wheat_meta)

setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/triticum/04_Kallisto_allreads_symlinked/")
files_t <- file.path(wheat_meta$ID, "abundance.h5")
names(files_t) <- paste0(wheat_meta$ID)
head(files_t)
all(file.exists(files_t))

# use tx2gene data.frame made earlier as txdb_hord and tx2gene_hord
# create count table
txi_trit <- tximport(files_t, type = "kallisto", tx2gene = tx2gene_trit)
dim(txi_trit$counts)
head(txi_trit$counts)

# save count table
setwd("/nfs/pgsb/projects/comparative_triticeae/phenotype/flower_development/refsets/triticum/")
write.table(txi_trit, file = "wheat_count_table_merged.txt", append = FALSE, quote = FALSE, sep = "\t", dec = ".",
            row.names = TRUE, col.names = TRUE)
saveRDS(txi_trit, file = "wheat_count_table_merged.rds")

sessionInfo()
